requirements.txt
psutil~=5.9.0
py-cpuinfo~=9.0.0
gputil~=1.4.0
keyboard~=0.13.5
rich~=13.9.4
pyamdgpuinfo